Refactor metadata imports and enhance type definitions across various pages
