+++ app/apps/public-web/.eslintrc.cjs
@@ -5,4 +5,12 @@ module.exports = {
     tsconfigRootDir: __dirname,
     project: ["./tsconfig.eslint.json"],
   },
+  settings: {
+    "import/resolver": {
+      typescript: {
+        alwaysTryTypes: true,
+        project: [`${__dirname}/tsconfig.json`, `${__dirname}/tsconfig.eslint.json`],
+      },
+    },
+  },
 };
