+++ app/apps/public-web/app/[locale]/(public)/forgot-password/page.tsx
@@ -1,11 +1,9 @@
 import type { Metadata } from "next";
 import type { ReactElement } from "react";
 
-import type { MetadataParams } from "../../../[locale]/layout";
+import { getPageMetadata, type MetadataParams } from "@/i18n/metadata";
 
 import { ForgotPasswordContent } from "./ForgotPasswordContent";
-
-import { getPageMetadata } from "@/i18n/metadata";
 
 export async function generateMetadata({ params }: MetadataParams): Promise<Metadata> {
   const { locale } = await params;
