+++ app/apps/public-web/app/[locale]/(public)/reset-password/page.tsx
@@ -1,11 +1,9 @@
 import type { Metadata } from "next";
 import type { ReactElement } from "react";
 
-import type { MetadataParams } from "../../../[locale]/layout";
+import { getPageMetadata, type MetadataParams } from "@/i18n/metadata";
 
 import { ResetPasswordContent } from "./ResetPasswordContent";
-
-import { getPageMetadata } from "@/i18n/metadata";
 
 export async function generateMetadata({ params }: MetadataParams): Promise<Metadata> {
   const { locale } = await params;
